requirements.txt
build
flufl.lock # for cartopy mpl tests
pytest
requests_mock
responses # mocking urllib3 requests
setuptools
wheel
